+++ CHANGELOG.md
@@ -5,6 +5,7 @@ AppEngine version, listed here to ease deployment and troubleshooting.
  * Bumped runtimeVersion to `2022.06.02`.
  * Upgraded stable Dart analysis SDK to `2.17.3`.
  * Upgraded preview Flutter analysis SDK to `3.1.0`.
+ * NOTE: Started to use and backfill `Package.isBlocked`.
 
 ## `20220530t115600-all`
  * Bumped runtimeVersion to `2022.05.23`.